requirements.txt
Django==4.1.4
django-location-field==2.1.0
djangorestframework==3.14.0
django-ajax-selects==2.1.0
django-embed-video==1.4.8
django-role-permissions==3.1.1
django-debug-toolbar==3.2
django-adminplus==0.5
django_filter==22.1
django-redis==4.12.1
django-cachalot==2.5.2
django-ipware==3.0.2
django-allauth==0.51.0
django-etc>=1.2.0
django-crispy-forms==1.14.0
django-simple-pwa==2.2.1
django_gulp==4.1.0
django-import-export==3.0.2
gunicorn==20.1.0
gevent==22.10.2
paytmchecksum==1.7.0
pillow==9.3.0
pycryptodome==3.10.1
whitenoise==5.2.0
psycopg2-binary==2.9.5
geopy==2.1.0
pydantic==1.10.2
python-dotenv==0.17.1
python-gettext==4.1
meilisearch>=0.21.0
requests==2.25.1
XlsxWriter==3.0.3
openpyxl==3.0.10
